requirements_wecom.txt
# 企业微信集成所需依赖
requests>=2.28.0
pycryptodome>=3.15.0
websockets>=10.3
tornado>=6.2
pymysql>=1.0.2
jieba>=0.42.1
gensim>=4.2.0